+++ ChangeLog.md
@@ -11,6 +11,7 @@ Eventum Issue Tracking System
 - remove support for calling deprecated handleAssignment workflow method (d16ea3a)
 - add Message-Id column to mail queue table (#140, @glensc)
 - cleanup deprecated "APP_LOCAL_PATH" from the template search path
+- drop unused Misc::getInput
 
 2016-04-28, Version [3.1.0]
 ----------------------------